feat: refactor rule create
