+++ ui/react/components/Dialog/Section/Splits/index.js
@@ -63,7 +63,6 @@ const Splits = React.createClass({
     },
     defaultSelected(origin, selected) {
         let result = [];
-        debugger;
         selected.forEach(function(row) {
             result.push(origin.filter(function(r) {
                 return r.key === row.key;
@@ -74,7 +73,6 @@ const Splits = React.createClass({
     },
     createSplitRows() {
         var self = this;
-        debugger;
 
         return this.props.data.map((split, index) => (
             <div key={index} style={{marginBottom: '20px'}}>
