--- ui/react/containers/Tabs/Index/actionTypes.js
@@ -1,3 +1,0 @@
-export const FETCH_RULES = Symbol('FETCH_RULES');
-export const FETCH_NOMENCLATURES = Symbol('FETCH_NOMENCLATURES');
-export const CREATE_RULE = Symbol('CREATE_RULE');
