+++ ui/react/containers/Tabs/Source/reducer.js
@@ -1,6 +1,7 @@
 import { Map, List } from 'immutable';
 
 import * as actionTypes from './actionTypes';
+import { RESET_RULE_STATE } from '../../../pages/Detail/actionTypes';
 
 const defaultState = Map({
     fields: Map({
@@ -31,6 +32,8 @@ export const ruleSourceTabReducer = (state = defaultState, action) => {
                 ['fields', 'properties', action.params.propertyId, action.params.field],
                 action.params.newValue
             );
+        case RESET_RULE_STATE:
+            return defaultState;
         default:
             return state;
     }
