+++ ui/react/index.js
@@ -1,9 +1,7 @@
 import React, { PropTypes } from 'react';
 import { Route, IndexRoute } from 'react-router';
 
-import { Main } from './pages';
-import Rewrite from './pages/Rewrite';
-import Tab from './containers/Tabs/Index';
+import { Main, Detail } from './pages';
 import { getRoute } from 'ut-front/react/routerHelper';
 import registerRoutes from './registerRoutes';
 
@@ -33,11 +31,8 @@ export const UtRuleRoutes = (config) => {
             <Route path={getRoute('ut-rule:home')}>
                 <IndexRoute component={Main} />
             </Route>
-            <Route path={getRoute('ut-rule:rewrite')}>
-                <IndexRoute component={Rewrite} />
-            </Route>
-            <Route path={getRoute('ut-rule:detail')}>
-                <IndexRoute component={Tab} />
+            <Route path={getRoute('ut-rule:create')}>
+                <IndexRoute component={Detail} />
             </Route>
         </Route>
     );
