+++ ui/react/pages/Detail/actionTypes.js
@@ -1,0 +1,4 @@
+export const FETCH_RULES = Symbol('FETCH_RULES');
+export const FETCH_NOMENCLATURES = Symbol('FETCH_NOMENCLATURES');
+export const CREATE_RULE = Symbol('CREATE_RULE');
+export const RESET_RULE_STATE = Symbol('RESET_RULE_STATE');
