--- ui/react/pages/Rewrite/actionTypes.js
@@ -1,2 +1,0 @@
-export const FETCH_RULES = Symbol('FETCH_RULES');
-export const FETCH_NOMENCLATURES = Symbol('FETCH_NOMENCLATURES');
