--- ui/react/pages/Rewrite/actions.js
@@ -1,13 +1,0 @@
-import * as actionTypes from './actionTypes';
-
-export const fetchRules = (params) => ({
-    type: actionTypes.FETCH_RULES,
-    method: 'db/rule.rule.fetch',
-    params: params || {}
-});
-
-export const fetchNomenclatures = (params) => ({
-    type: actionTypes.FETCH_NOMENCLATURES,
-    method: 'rule.item.fetch',
-    params: params || {}
-});
