--- ui/react/pages/Rewrite/helpers.js
@@ -1,15 +1,0 @@
-export const formatRuleItems = (items) => {
-    let formattedPayload = {};
-
-    items.map(item => {
-        if (!formattedPayload[item.type]) {
-            formattedPayload[item.type] = [];
-        }
-        formattedPayload[item.type].push({
-            key: item.value,
-            name: item.display
-        });
-    });
-
-    return formattedPayload;
-};
