+++ ui/react/pages/index.js
@@ -1,3 +1,4 @@
 import Main from './Main';
+import Detail from './Detail';
 
-export { Main };
+export { Main, Detail };
