+++ ui/react/registerRoutes.js
@@ -2,10 +2,9 @@ import { registerRoute } from 'ut-front/react/routerHelper';
 
 export default () => {
     let mainRoute = 'ut-rule:home';
-    let rewriteRoute = 'ut-rule:rewrite';
-    let createRule = 'ut-rule:detail';
+    // let rewriteRoute = 'ut-rule:rewrite';
+    let createRule = 'ut-rule:create';
     registerRoute(mainRoute).path('/rule');
-    registerRoute(rewriteRoute).path('/rule/rewrite');
     registerRoute(createRule).path('/rule/create');
     return mainRoute;
 };
